FFT processing branch is manually merged cause development needs to be paused for now and we want to avoid messy merge when development continues. Calling of FFT processing and drawing is commented out and TODO comments are added for easier reference.

+++ app/src/main/java/com/backyardbrains/utils/CircularShortBuffer.java
@@ -42,7 +42,7 @@ public class CircularShortBuffer {
     /**
      * Gets a single short return or -1 if no data is available.
      */
-    public synchronized int get() {
+    public synchronized short get() {
         if (available == 0) {
             return -1;
         }
@@ -150,7 +150,7 @@ public class CircularShortBuffer {
     /**
      * Return the first short a <b>get</b> would return or -1 if no data is available.
      */
-    public synchronized int peek() {
+    public synchronized short peek() {
         return available > 0 ? buffer[idxGet] : -1;
     }
